Fix compile errors for .NET 4.5 portable project
Mostly related to reflection. Solved using compile conditions on
PORTABLE45.

WinStore introduced a new reflection API with TypeInfo and newer PCL
profiles use this instead, but the existing Portable project is the
older profile 136 that does not support this, so we need to
distinguish between PORTABLE (profile 136) and PORTABLE45 (profile 259).

The new reflection API is only available on .NET 4.5 and later, so as
long as .NET 4.0 support is desired we will need to distinguish these
two. An improvement would be to rename the PORTABLE condition to
PORTABLE40.

Breaking change in ExceptionFormatter.cs:
Thread information is not available in profile 259 as
far as I can see. I've set the string to "N/A" on the thread ID for
now. Either we must inject a function to retrieve the thread ID using
a platform specific API or find a way to access it from PCL profile 259.